Filter sort + Test Beer + MySql storage on Docker

+++ src/main/java/ch/hearc/jee2024/project/IOC/OrderController.java
@@ -1,6 +1,7 @@
 package ch.hearc.jee2024.project.IOC;
 
 import ch.hearc.jee2024.project.ServiceOrder.OrderService;
+import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -57,14 +58,29 @@ public class OrderController {
     }
 
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable Long id) {
-        orderService.deleteOrder(id);
-        return ResponseEntity.noContent().build();
+    @DeleteMapping("/{id}/checkout")
+    public ResponseEntity<Double> deleteOrder(@PathVariable Long id) {
+        Order order = orderService.getOrderById(id).orElse(null);
+        if(order == null) {
+            return ResponseEntity.notFound().build();
+        }else {
+            orderService.deleteOrder(id);
+            return ResponseEntity.ok(order.getTotalPrice());
+        }
     }
 
+
     @GetMapping
-    public ResponseEntity<List<Order>> getAllOrders() {
-        return ResponseEntity.ok(orderService.getAllOrders());
+    public ResponseEntity<Page<Order>> getOrders(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size,
+            @RequestParam(defaultValue = "dateCreated") String sortBy,
+            @RequestParam(defaultValue = "desc") String direction) {
+
+        Page<Order> orders = orderService.getOrders(page, size, sortBy, direction);
+        return ResponseEntity.ok(orders);
     }
+
+
 }
+
